radio buttons for crash simulation

+++ server.js
@@ -221,6 +221,8 @@ app.get('/api/data', async (req, res) => {
 
 // create
 app.post('/api/create', async (req, res) => {
+    const crashNode = req.headers['crash-node'];
+
     const {tconst, primary_title, original_title, title_type, is_adult, runtime_minutes, year} = req.body;
 
     const values = [tconst, primary_title, original_title, title_type, is_adult, runtime_minutes, year];
@@ -244,6 +246,11 @@ app.post('/api/create', async (req, res) => {
             let query, currentValues;
 
             if (target === 'Master') {
+
+                if (crashNode === 'master-crash') {
+                    throw new Error("Simulated Master crash before INSERT operation.");
+                }
+
                 query = `INSERT INTO metadata${tableSuffix[target]} (tconst, primary_title, original_title, title_type, is_adult, runtime_minutes, year) VALUES (?, ?, ?, ?, ?, ?, ?)`;
                 
                 const [result] = await pools['Master'].query(query, values);
@@ -252,7 +259,6 @@ app.post('/api/create', async (req, res) => {
                 generatedId = result.insertId;
                 console.log(`Master generated ID: ${generatedId}`);
 
-                // log the transaction before insertion into slaves
                 logId = await logTransaction(currentNode, 'INSERT', `metadata${tableSuffix['Master']}`, generatedId, null, newValues, targetNodes);
 
                 successfulNodes.push(target);
@@ -278,6 +284,13 @@ app.post('/api/create', async (req, res) => {
                         console.error("CRITICAL: One slave is also down! Could not query Slaves for Max ID.", idErr);
                         throw new Error("ID Generation failed. Cluster unavailable.");
                     }
+                }
+
+                if (target === 'OldSlave' && crashNode === 'oldslave-crash') {
+                    throw new Error("Simulated OldSlave crash during INSERT operation replication.");
+                } 
+                if (target === 'NewSlave' && crashNode === 'newslave-crash') {
+                    throw new Error("Simulated NewSlave crash during INSERT operation replication.");
                 }
 
                 query = `INSERT INTO metadata${tableSuffix[target]} 
@@ -369,6 +382,8 @@ app.get('/api/edit', async (req, res) => {
 
 // POST FOR EDIT
 app.post('/api/update', async (req, res) => {
+    const crashNode = req.headers['crash-node'];
+
     console.log("Received update request:", req.body);
 
     const {
@@ -426,6 +441,12 @@ app.post('/api/update', async (req, res) => {
     if (isMigration) {
         console.log(`Migration detected: Moving from ${sourceSlave} to ${destSlave}`);
         try {
+            const isSourceCrash = (sourceSlave === 'OldSlave' && crashNode === 'oldslave-crash') || (sourceSlave === 'NewSlave' && crashNode === 'newslave-crash');
+
+            if (isSourceCrash) {
+                throw new Error(`Simulated ${sourceSlave} crash during DELETE operation of migration.`);
+            }
+
             console.log(`Deleting from old node ${sourceSlave}...`);
             const delQuery = `DELETE FROM metadata${tableSuffix[sourceSlave]} WHERE metadata_key = ?`;
             await pools[sourceSlave].query(delQuery, [metadata_key]);
@@ -465,6 +486,17 @@ app.post('/api/update', async (req, res) => {
                             WHERE metadata_key = ?`;
                 params = updateParams;
             }
+
+            if(target === 'Master' && crashNode === 'master-crash') {
+                throw new Error(`Simulated Master crash during UPDATE operation.`);
+            }
+            if(target === 'OldSlave' && crashNode === 'oldslave-crash') {
+                throw new Error(`Simulated OldSlave crash during UPDATE/MIGRATION INSERT operation.`);
+            }
+            if(target === 'NewSlave' && crashNode === 'newslave-crash') {
+                throw new Error(`Simulated NewSlave crash during UPDATE/MIGRATION INSERT operation.`);
+            }
+
             await pools[target].query(query, params);
             successfulNodes.push(target);
             successCount++;
@@ -609,7 +641,7 @@ app.post('/api/recover', async (req, res) => {
                                     console.error(`Failed to delete migrated record from old slave ${oldSlave}:`, delErr);
                                 }
                             }
-                            
+
                             successfulNodes.push(target);
                             successCount++;
                         }
